Extract Issue_Detector from AI_Formatter for reuse
Move issue-detection logic (detect_issues, build_issue_description,
run_heuristic_checks, find_section_by_id) from AI_Formatter into a
standalone Issue_Detector class so the Abilities API provider and
other consumers can detect issues without formatting.

AI_Formatter now delegates to Issue_Detector, preserving identical
output and filter behavior.
